Moved all ASM hooks into one class
