+++ src/main/java/com/mrcrayfish/controllable/client/ControllerInput.java
@@ -824,79 +824,4 @@ public class ControllerInput
             }, "mouseReleased event handler", screen.getClass().getCanonicalName());
         }
     }
-
-    /**
-     * Used in order to fix block breaking progress. This method is linked via ASM.
-     */
-    public static boolean isLeftClicking()
-    {
-        Minecraft mc = Minecraft.getInstance();
-        boolean isLeftClicking = mc.gameSettings.keyBindAttack.isKeyDown();
-        Controller controller = Controllable.getController();
-        if(controller != null)
-        {
-            if(ButtonBindings.ATTACK.isButtonDown())
-            {
-                isLeftClicking = true;
-            }
-        }
-        boolean usingVirtualMouse = (Controllable.getOptions().isVirtualMouse() && Controllable.getInput().getLastUse() > 0);
-        return mc.currentScreen == null && isLeftClicking && (mc.mouseHelper.isMouseGrabbed() || usingVirtualMouse);
-    }
-
-    /**
-     * Used in order to fix actions like eating or pulling bow back. This method is linked via ASM.
-     */
-    @SuppressWarnings("unused")
-    public static boolean isRightClicking()
-    {
-        Minecraft mc = Minecraft.getInstance();
-        boolean isRightClicking = mc.gameSettings.keyBindUseItem.isKeyDown();
-        Controller controller = Controllable.getController();
-        if(controller != null)
-        {
-            if(ButtonBindings.USE_ITEM.isButtonDown())
-            {
-                isRightClicking = true;
-            }
-        }
-        return isRightClicking;
-    }
-
-    /**
-     * Used in order to fix the quick move check in inventories. This method is linked via ASM.
-     */
-    @SuppressWarnings("unused")
-    public static boolean canQuickMove()
-    {
-        boolean canQuickMove = InputMappings.isKeyDown(Minecraft.getInstance().mainWindow.getHandle(), GLFW.GLFW_KEY_LEFT_SHIFT) || InputMappings.isKeyDown(Minecraft.getInstance().mainWindow.getHandle(), GLFW.GLFW_KEY_RIGHT_SHIFT);
-        Controller controller = Controllable.getController();
-        if(controller != null)
-        {
-            if(ButtonBindings.QUICK_MOVE.isButtonDown())
-            {
-                canQuickMove = true;
-            }
-        }
-        return canQuickMove;
-    }
-
-    /**
-     * Allows the player list to be shown. This method is linked via ASM.
-     */
-    @SuppressWarnings("unused")
-    public static boolean canShowPlayerList()
-    {
-        Minecraft mc = Minecraft.getInstance();
-        boolean canShowPlayerList = mc.gameSettings.keyBindPlayerList.isKeyDown();
-        Controller controller = Controllable.getController();
-        if(controller != null)
-        {
-            if(ButtonBindings.PLAYER_LIST.isButtonDown())
-            {
-                canShowPlayerList = true;
-            }
-        }
-        return canShowPlayerList;
-    }
 }
